+++ examples/resources/elasticstack_elasticsearch_transform/resource.tf
@@ -8,6 +8,16 @@ resource "elasticstack_elasticsearch_transform" "transform_with_pivot" {
 
   destination {
     index = "destination_index_for_transform"
+
+    aliases {
+      alias            = "test_alias_1"
+      move_on_creation = true
+    }
+
+    aliases {
+      alias            = "test_alias_2"
+      move_on_creation = false
+    }
   }
 
   pivot = jsonencode({